+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - Parsing cluster-wide config with empty roles list
+- Parsing 'number' args on calling admin functions
 
 ## [2.5.0] - 2020-12-29