+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.6.0] - 2021-01-27
+
 ### Added
 
 - `--conn, -c` option for `cartridge admin` command to specify instance address